Fix logo URL
The new `dat-data.com` no longer has the dat logo at the URL that it used to be at. As a result, launching a new Heroku instance displays a broken image icon on the Heroku dashboard. This change points Heroku at the new URL, although with perhaps a lower-quality image than might look ideal on Heroku's site.